refactor: consolidate agent module splits (project, ops_paths) [OPE-424, OPE-426, OPE-427]
Split monolithic files into focused submodules:
- cli cmd/project.rs → cmd/project/ (OPE-426)
- projects project.rs → project/ (OPE-427)
- web ops_paths.rs → ops_paths/ (OPE-424)
