+++ crates/opengoose-projects/src/project/mod.rs
@@ -1,0 +1,9 @@
+mod context;
+mod definition;
+mod path;
+
+#[cfg(test)]
+mod tests;
+
+pub use context::ProjectContext;
+pub use definition::{ProjectDefinition, ProjectSettings};
