+++ crates/opengoose-projects/src/project/path.rs
@@ -1,0 +1,16 @@
+use std::path::PathBuf;
+
+/// Expand a leading `~` to the home directory.
+pub(super) fn expand_tilde(path: &str) -> PathBuf {
+    if let Some(rest) = path.strip_prefix("~/") {
+        if let Some(home) = dirs::home_dir() {
+            return home.join(rest);
+        }
+    } else if path == "~"
+        && let Some(home) = dirs::home_dir()
+    {
+        return home;
+    }
+
+    PathBuf::from(path)
+}
